Add test data for ridges_circle_pass in CSV format
This commit introduces a new CSV file containing test data for the ridges_circle_pass functionality. The data includes two columns: 'tau' and 'f', with a total of 4807 entries. This dataset will be used for validating the performance and accuracy of the related algorithms in the cosmology black holes module.